Remove a11y check and refactor components
