+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "svelte.plugin.svelte.compilerWarnings": {
+    "a11y-click-events-have-key-events": "ignore",
+    "a11y-no-static-element-interactions": "ignore"
+  }
+}
