+++ src/components/UnitsList.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import BuilderStore from '$builder/store'
-  import BuilderItem from './items/BuilderItem.svelte'
+  import Unit from './items/Unit.svelte'
 </script>
 
 <div class="w-1/3 divide-y divide-gray-200 text-center">
@@ -11,7 +11,5 @@
     <div class="w-1/5">Points</div>
     <div class="w-1/5">Min/Max</div>
   </div>
-  {#each $BuilderStore.units as unit}
-    <BuilderItem {unit} />
-  {/each}
+  {#each $BuilderStore.units as unit} <Unit {unit} /> {/each}
 </div>
