+++ src/routes/Builder.svelte
@@ -5,7 +5,7 @@
   import BuilderStore from '$builder/store'
 
   import SchemaList from '$root/src/components/SchemaList.svelte'
-  import BuilderList from '$root/src/components/BuilderList.svelte'
+  import UnitsList from '$root/src/components/UnitsList.svelte'
   import BuilderInfo from '$root/src/components/BuilderInfo.svelte'
 
   export let factionFile: string
@@ -35,7 +35,7 @@
   <section class="flex justify-evenly items-start">
     <SchemaList armySchema={armyData} />
     <BuilderInfo armyName={armyData.name} />
-    <BuilderList />
+    <UnitsList />
   </section>
 {:catch error}
   <Link to="/">Return to homepage</Link>
